test: Rename AmazonChimeSDKCuckooTests to AmazonChimeSDKTests
With Mockingbird removed there is a single unit-test target, and naming
it after its mocking library no longer makes sense. Rename the target,
scheme, and source folder to AmazonChimeSDKTests, and merge the two test
source trees into one.

- All 31 files from AmazonChimeSDKCuckooTests/ move into
  AmazonChimeSDKTests/ at the same relative paths; the duplicate Xcode
  group hierarchy is merged so the navigator shows a single tree.
- The stale AmazonChimeSDKTests/common/CommonTestCase.swift copy is
  deleted; its build-file entry had belonged to no target since the
  Mockingbird target was removed.
- AmazonChimeSDKCuckooTests.xcscheme becomes AmazonChimeSDKTests.xcscheme;
  AmazonChimeSDK.xcscheme's TestAction follows. Coverage stays enabled.
- Cuckoofile.toml output path, the .swiftlint.yml exclusion, and the
  codecov.yml scheme and -skip-testing references are updated.

Verified: 572 tests across 90 suites pass, and AmazonChimeSDK.framework
line coverage is unchanged at 5604/8189 (68.43%).
