+++ AmazonChimeSDK/AmazonChimeSDKTests/ingestion/NetworkConnectionTypeTests.swift
